Rename files JC-718
